Dynamically adjust JRE memory limits
This is required because the JRE by default limits the heap up to 25%
of available memory. A pod must have a memory limit and the
application is expected to use this limit, so 25% doesn't make sense.

The added code will allow for heap size up to 70% of the limit if the
limit is below 1Gi, up to 75% if the limit is below 1.5Gi, and up to
85% for rest of cases.

We also set the default memory limit to 1Gi since this means a heap of
up to 700Mi. Before, although we set the limit to 2Gi, the heap could
go up to 500Mi.

[OPS-2432]

Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>